Replaced char by byte for enum

+++ src/com/mideas/rpg/v2/game/item/weapon/WeaponType.java
@@ -2,29 +2,29 @@ package com.mideas.rpg.v2.game.item.weapon;
 
 public enum WeaponType {
 
-	ONEHANDEDAXE(0),
-	TWOHANDEDAXE(1),
-	ONEHANDEDSWORD(2),
-	TWOHANDEDSWORD(3),
-	ONEHANDEDMACE(4),
-	TWOHANDEDMACE(5),
-	POLEARM(6),
-	STAFF(7),
-	DAGGER(8),
-	FISTWEAPON(9),
-	BOW(10),
-	CROSSBOW(11),
-	GUN(12),
-	THROWN(13),
-	WAND(14);
+	ONEHANDEDAXE((byte)0),
+	TWOHANDEDAXE((byte)1),
+	ONEHANDEDSWORD((byte)2),
+	TWOHANDEDSWORD((byte)3),
+	ONEHANDEDMACE((byte)4),
+	TWOHANDEDMACE((byte)5),
+	POLEARM((byte)6),
+	STAFF((byte)7),
+	DAGGER((byte)8),
+	FISTWEAPON((byte)9),
+	BOW((byte)10),
+	CROSSBOW((byte)11),
+	GUN((byte)12),
+	THROWN((byte)13),
+	WAND((byte)14);
 	
-	private int value;
+	private byte value;
 	
-	private WeaponType(int value) {
+	private WeaponType(byte value) {
 		this.value = value;
 	}
 	
-	public int getValue() {
+	public byte getValue() {
 		return this.value;
 	}
 }
